+++ src/Task1.cc
@@ -57,6 +57,8 @@ Task1::Task1(const sdf::ElementPtr &_sdf) : Task(_sdf)
   std::unique_ptr<Task1CP4> cp4(new Task1CP4(cp4Elem));
   this->checkpoints.push_back(std::move(cp4));
 
+  this->logFilter = "satellite_dish*|valkyrie*";
+
   gzmsg << "Task [1] created" << std::endl;
 }
 
@@ -73,13 +75,13 @@ bool Task1CP1::Check()
 }
 
 /////////////////////////////////////////////////
-void Task1CP1::Restart()
+void Task1CP1::Restart(const common::Time &_penalty)
 {
   // This is the 1st CP of the task: reharness back at start box
   HarnessManager::Instance()->NewGoal(
       ignition::math::Pose3d(0, 0, 1.257, 0, 0, 0));
 
-  Checkpoint::Restart();
+  Checkpoint::Restart(_penalty);
 }
 
 /////////////////////////////////////////////////
@@ -94,6 +96,8 @@ bool Task1CP2::Check()
   // First time
   if (!this->satelliteRosSub && !this->oneAxisDone)
   {
+    this->Start();
+
     // Subscribe to satellite msgs
     this->rosNode.reset(new ros::NodeHandle());
     this->satelliteRosSub = this->rosNode->subscribe(